@@ -1,14 +1,15 @@
 {
   "name": "sirv-cli",
-  "version": "0.4.6",
+  "version": "1.0.0-next.9",
   "description": "A lightweight CLI program to serve static sites~!",
   "repository": "lukeed/sirv",
+  "main": "index.js",
   "license": "MIT",
   "files": [
     "*.js"
   ],
   "bin": {
-    "sirv": "index.js"
+    "sirv": "bin.js"
   },
   "author": {
     "name": "Luke Edwards",
@@ -16,15 +17,16 @@
     "url": "https://lukeed.com"
   },
   "engines": {
-    "node": ">= 6"
+    "node": ">= 10"
   },
   "dependencies": {
     "console-clear": "^1.1.0",
     "get-port": "^3.2.0",
     "kleur": "^3.0.0",
     "local-access": "^1.0.1",
-    "sade": "^1.4.0",
-    "sirv": "^0.4.6",
+    "sade": "^1.6.0",
+    "semiver": "^1.0.0",
+    "sirv": "^1.0.0-next.9",
     "tinydate": "^1.0.0"
   }
 }
