@@ -1,11 +1,15 @@
 {
   "name": "sirv",
-  "version": "0.4.6",
+  "version": "1.0.0-next.9",
   "description": "The optimized & lightweight middleware for serving requests to static assets",
   "repository": "lukeed/sirv",
+  "module": "build.mjs",
+  "types": "sirv.d.ts",
+  "main": "build.js",
   "license": "MIT",
   "files": [
-    "*.js"
+    "build.*",
+    "sirv.d.ts"
   ],
   "author": {
     "name": "Luke Edwards",
@@ -13,10 +17,11 @@
     "url": "https://lukeed.com"
   },
   "engines": {
-    "node": ">= 6"
+    "node": ">= 10"
   },
   "dependencies": {
-    "@polka/url": "^0.5.0",
-    "mime": "^2.3.1"
+    "@polka/url": "^1.0.0-next.9",
+    "mime": "^2.3.1",
+    "totalist": "^1.0.0"
   }
 }
